revisi laporan dll

+++ app/Http/Controllers/DivisiController.php
@@ -883,16 +883,20 @@ class DivisiController extends Controller
 
     public function proseseditlaporan($id, Request $request)
     {
+        if ($request->file('path') != null){
         $lama = Laporan::find($id);
-        File::delete('file/laporan-mhs-revisi/' . $lama->path);
+        File::delete('file/laporan-mhs/' . $lama->path);
         $file = $request->file('path');
         $nama_file = $file->getClientOriginalName();
-        $tujuan_upload = 'file/laporan-mhs-revisi/';
+        $tujuan_upload = 'file/laporan-mhs/';
         $file->move($tujuan_upload, $nama_file);
+        }else{
+            $nama_file = null;
+        }
         DB::table('laporans')->where('id', $id)
             ->update([
                 'nama_pembimbing_hcd'=> $request->nama_pembimbing_hcd,
-                'path_revisi'=> $nama_file,
+                'path'=> $nama_file,
                 'revisi' => $request->revisi
             ]);
 
@@ -904,7 +908,6 @@ class DivisiController extends Controller
         $laporanmhs = DB::table('laporans')->find($id);
         // Hapus di file storage
         File::delete('file/laporan-mhs/' . $laporanmhs->path);
-        File::delete('file/laporan-mhs-revisi/' . $laporanmhs->path_revisi);
         // Hapus di database
         DB::table('laporans')
             ->where('id', $laporanmhs->id)
@@ -925,17 +928,20 @@ class DivisiController extends Controller
     }
     public function proseseditlaporansmk($id, Request $request)
     {
+        if ($request->file('path') != null){
         $lama = LaporanSmk::find($id);
         File::delete('file/laporan-smk-revisi/' . $lama->path);
-
-        $file = $request->file('path_revisi');
+        $file = $request->file('path');
         $nama_file = $file->getClientOriginalName();
-        $tujuan_upload = 'file/laporan-smk-revisi/';
+        $tujuan_upload = 'file/laporan-smk/';
         $file->move($tujuan_upload, $nama_file);
+        }else{
+            $nama_file = null;
+        }
         DB::table('laporans_smk')->where('id', $id)
             ->update([
                 'nama_pembimbing_hcd'=> $request->nama_pembimbing_hcd,
-                'path_revisi'=> $nama_file,
+                'path'=> $nama_file,
                 'revisi' => $request->revisi
             ]);
 
@@ -947,7 +953,6 @@ class DivisiController extends Controller
         $laporansmk = DB::table('laporans_smk')->find($id);
         // Hapus di file storage
         File::delete('file/laporan-smk/' . $laporansmk->path);
-        File::delete('file/laporan-smk-revisi/' . $laporansmk->path_revisi);
         // Hapus di database
         DB::table('laporans_smk')
             ->where('id', $laporansmk->id)
@@ -2858,19 +2863,23 @@ class DivisiController extends Controller
 
     public function proseseditlaporan_divisi($id, Request $request)
     {
+        if ($request->file('path') != null){
         $lama = Laporan::find($id);
-        File::delete('file/laporan-mhs-revisi/' . $lama->path);
-        $file = $request->file('path_revisi');
+        File::delete('file/laporan-mhs/' . $lama->path);
+        $file = $request->file('path');
         $nama_file = $file->getClientOriginalName();
-        $tujuan_upload = 'file/laporan-mhs-revisi/';
+        $tujuan_upload = 'file/laporan-mhs/';
         $file->move($tujuan_upload, $nama_file);
+        }else{
+        $nama_file = null;
+        }
         DB::table('laporans')->where('id', $id)
             ->update([
                 'nama_pembimbing_lapangan'=> $request->nama_pembimbing_lapangan,
-                'path_revisi'=> $nama_file,
+                'path'=> $nama_file,
                 'revisi_divisi' => $request->revisi_divisi
             ]);
-
+       
         session()->flash('succes', 'Data anda berhasil di update');
         return redirect('/laporan-divisi');
     }
@@ -2886,18 +2895,20 @@ class DivisiController extends Controller
     }
     public function proseseditlaporansmk_divisi($id, Request $request)
     {
-        
+        if ($request->file('path') != null){
         $lama = LaporanSmk::find($id);
-        File::delete('file/laporan-smk-revisi/' . $lama->path);
-
-        $file = $request->file('path_revisi');
+        File::delete('file/laporan-smk/' . $lama->path);
+        $file = $request->file('path');
         $nama_file = $file->getClientOriginalName();
-        $tujuan_upload = 'file/laporan-smk-revisi/';
+        $tujuan_upload = 'file/laporan-smk';
         $file->move($tujuan_upload, $nama_file);
+        }else{
+            $nama_file = null;
+        }
         DB::table('laporans_smk')->where('id', $id)
             ->update([
                 'nama_pembimbing_lapangan'=> $request->nama_pembimbing_lapangan,
-                'path_revisi'=> $nama_file,
+                'path'=> $nama_file,
                 'revisi_divisi' => $request->revisi_divisi
             ]);
